backend/microservicios_reservas/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
flask_jwt_extended==4.6.0